Inspect all configs in manage-projects
We were skipping the setting of retired project acls in manage-projects
because we were iterating over projects that weren't retired. In
manage-projects we don't want to skip over retired projects so that we
can retire the projects properly.

Address this by adding a new registry property that returns all projects
retired or not then iterate over that in manage-projects.

Change-Id: I7aa89b53696fb0fca262fc7068e5e2ad2b7ad8fc